Add testmods for alternative cloud optics schemes
Add testmods to allow testing alternative cloud optics configurations.
This allows testing with configurations using the Slingo and Ebert-Curry
cloud optics schemes, and also testing using the "oldcldoptics" flag.
These schemes are used for the MMF single-moment microphysics
configurations, so testing these is important as we refactor to make
sure we do not change answers.