Remove defunct and misleading oldcldoptics flag
Remove oldcldoptics flag. This flag was not in use by any of our
configurations, and was also misleading as it resulted in different
behavior for SW and LW optics. For SW, it would force RRTMG to call the
Slingo and Ebert-Curry optics schemes, but for LW, it would result in
calling these routines only for diagnostics, but then overwrite them
with calls to newer optics schemes.